move player area constants (and future 'is inside?' check) to a lib object instead of inside setup
